+++ pmd-core/src/main/java/net/sourceforge/pmd/cpd/CpdCapableLanguage.java
@@ -16,7 +16,7 @@ public interface CpdCapableLanguage extends Language {
 
 
     /**
-     * Create a new {@link Tokenizer} for this language, given
+     * Create a new {@link CpdLexer} for this language, given
      * a property bundle with configuration. The bundle was created by
      * this instance using {@link #newPropertyBundle()}. It can be assumed
      * that the bundle will never be mutated anymore, and this method
@@ -26,7 +26,7 @@ public interface CpdCapableLanguage extends Language {
      *
      * @return A new language processor
      */
-    default Tokenizer createCpdTokenizer(LanguagePropertyBundle bundle) {
-        return new AnyTokenizer();
+    default CpdLexer createCpdLexer(LanguagePropertyBundle bundle) {
+        return new AnyCpdLexer();
     }
 }