+++ pmd-core/src/main/java/net/sourceforge/pmd/cpd/Tokens.java
@@ -14,7 +14,7 @@ import org.checkerframework.checker.nullness.qual.NonNull;
 import org.checkerframework.checker.nullness.qual.Nullable;
 
 import net.sourceforge.pmd.annotation.InternalApi;
-import net.sourceforge.pmd.lang.ast.TokenMgrError;
+import net.sourceforge.pmd.lang.ast.LexException;
 import net.sourceforge.pmd.lang.document.FileId;
 import net.sourceforge.pmd.lang.document.TextDocument;
 
@@ -93,7 +93,7 @@ public class Tokens {
 
     /**
      * Creates a token factory to process the given file with
-     * {@link Tokenizer#tokenize(TextDocument, TokenFactory)}.
+     * {@link CpdLexer#tokenize(TextDocument, TokenFactory)}.
      * Tokens are accumulated in the {@link Tokens} parameter.
      *
      * @param file   Document for the file to process
@@ -117,8 +117,8 @@ public class Tokens {
             }
 
             @Override
-            public TokenMgrError makeLexException(int line, int column, String message, @Nullable Throwable cause) {
-                return new TokenMgrError(line, column, fileId, message, cause);
+            public LexException makeLexException(int line, int column, String message, @Nullable Throwable cause) {
+                return new LexException(line, column, fileId, message, cause);
             }
 
             @Override