+++ pmd-fortran/src/main/java/net/sourceforge/pmd/lang/fortran/FortranLanguageModule.java
@@ -4,8 +4,8 @@
 
 package net.sourceforge.pmd.lang.fortran;
 
-import net.sourceforge.pmd.cpd.AnyTokenizer;
-import net.sourceforge.pmd.cpd.Tokenizer;
+import net.sourceforge.pmd.cpd.AnyCpdLexer;
+import net.sourceforge.pmd.cpd.CpdLexer;
 import net.sourceforge.pmd.lang.LanguagePropertyBundle;
 import net.sourceforge.pmd.lang.LanguageRegistry;
 import net.sourceforge.pmd.lang.impl.CpdOnlyLanguageModuleBase;
@@ -30,7 +30,7 @@ public class FortranLanguageModule extends CpdOnlyLanguageModuleBase {
     }
 
     @Override
-    public Tokenizer createCpdTokenizer(LanguagePropertyBundle bundle) {
-        return new AnyTokenizer("!");
+    public CpdLexer createCpdLexer(LanguagePropertyBundle bundle) {
+        return new AnyCpdLexer("!");
     }
 }