+++ pmd-jsp/src/main/java/net/sourceforge/pmd/lang/jsp/JspLanguageModule.java
@@ -5,11 +5,11 @@
 package net.sourceforge.pmd.lang.jsp;
 
 import net.sourceforge.pmd.cpd.CpdCapableLanguage;
-import net.sourceforge.pmd.cpd.Tokenizer;
+import net.sourceforge.pmd.cpd.CpdLexer;
 import net.sourceforge.pmd.lang.LanguagePropertyBundle;
 import net.sourceforge.pmd.lang.LanguageRegistry;
 import net.sourceforge.pmd.lang.impl.SimpleLanguageModuleBase;
-import net.sourceforge.pmd.lang.jsp.cpd.JSPTokenizer;
+import net.sourceforge.pmd.lang.jsp.cpd.JspCpdLexer;
 
 /**
  * Created by christoferdutz on 20.09.14.
@@ -31,7 +31,7 @@ public class JspLanguageModule extends SimpleLanguageModuleBase implements CpdCa
     }
 
     @Override
-    public Tokenizer createCpdTokenizer(LanguagePropertyBundle bundle) {
-        return new JSPTokenizer();
+    public CpdLexer createCpdLexer(LanguagePropertyBundle bundle) {
+        return new JspCpdLexer();
     }
 }