+++ pmd-tsql/src/main/java/net/sourceforge/pmd/lang/tsql/TSqlLanguageModule.java
@@ -4,11 +4,11 @@
 
 package net.sourceforge.pmd.lang.tsql;
 
-import net.sourceforge.pmd.cpd.Tokenizer;
+import net.sourceforge.pmd.cpd.CpdLexer;
 import net.sourceforge.pmd.lang.LanguagePropertyBundle;
 import net.sourceforge.pmd.lang.LanguageRegistry;
 import net.sourceforge.pmd.lang.impl.CpdOnlyLanguageModuleBase;
-import net.sourceforge.pmd.lang.tsql.cpd.TSqlTokenizer;
+import net.sourceforge.pmd.lang.tsql.cpd.TSqlCpdLexer;
 
 /**
  * @author pguyot@kallisys.net
@@ -25,7 +25,7 @@ public class TSqlLanguageModule extends CpdOnlyLanguageModuleBase {
     }
 
     @Override
-    public Tokenizer createCpdTokenizer(LanguagePropertyBundle bundle) {
-        return new TSqlTokenizer();
+    public CpdLexer createCpdLexer(LanguagePropertyBundle bundle) {
+        return new TSqlCpdLexer();
     }
 }